Make DSE Audit log tuner bindable (with default binding to yaml). DSE Audit log is configured via log4j in 3.x and yaml since 4.x.
